nassh: harmonize script load ordering
We're inconsistent between the manifests, main page, and test page
when it comes to the order of loading scripts.  This causes issues
when we try to add new APIs and it works in one context but not in
the others.  Try to use the same order across all of them.

Change-Id: If43fe65e0789fe4400bd524251724ca9f49f1952
Reviewed-on: https://chromium-review.googlesource.com/c/1453976
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>