hterm: limit X10 encoding to 127 for now
When we moved the UTF-8 encoding to ArrayBuffers and out of hterm to
nassh ("hterm/nassh: switch to array buffers for sending tty data"),
we broke X10 mouse encoding.  Before, we weren't encoding its data in
UTF-8 but sending it directly on the wire, but now we're re-encoding
it into UTF-8 bytes.  For values over 127, we transmit 2 bytes when
it needs to be 1.  This is invalid UTF-8, and goes against our goals
in hterm to do Unicode everywhere.

While we figure out whether we want to support transmitting binary
data at all, limit the X10 encoding to 127 so we don't confuse remote
programs who expect the full byte.

Change-Id: Icefce607db7d45294159fb0271ed32817c7f4bc4
Reviewed-on: https://chromium-review.googlesource.com/c/1447104
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>