nassh: fix race with background page and sftp mounts
If the background page doesn't exist, the call to getBackgroundPage
will create it on the fly.  But it will return a handle as soon as
Chrome has created the window, not after it has finished loading.
We can end up racing and trying to use createSftpInstance before it
is ready leading to some objects being undefined.

Add a helper that will delay returning the background page until it
has fully initialized itself to avoid all of that.

Bug: chromium:930171
Change-Id: Id2ad6c14d1bc52ceb560d07497ed2443b8aa8b5e
Reviewed-on: https://chromium-review.googlesource.com/c/1462717
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>