nassh: sftp: change toByteArray helper to return typed arrays
The toByteArray helper returns an ArrayBuffer currently, but in some
contexts we want a typed array view.  So change toByteArray to return
a typed array and add a new toArrayBuffer to explicitly get the buffer
when needed.  This allows us to make a lot of mechanical changes (like
tests) before making functional changes to convert the underlying code
to use ArrayBuffers everywhere.

Change-Id: Ie0de18bd43902f72cf18b2d0a8877a7a2c3719e2
Reviewed-on: https://chromium-review.googlesource.com/c/1481988
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>