hterm: io: check for strings before ArrayBuffers
The way we currently handle sftp mounts is by:
- instantiating the terminal in the foreground page,
- getting a reference to the background page in the foreground page,
- passing a reference to the terminal directly to the background page,
- instantiating the SSH plugin in the background page,
- then the background page reads/writes plugin data to the terminal.

When the incoming data is an ArrayBuffer, it's instantiated using the
object from the background page.  But this isn't the same ArrayBuffer
as in the foreground page, so trying to use instanceof doesn't work.
For example:
> chrome.runtime.getBackgroundPage((bg) => window.bg = bg)
> ab1 = new ArrayBuffer()
> ab1 instanceof ArrayBuffer
true
> ab1 instanceof bg.ArrayBuffer
false
> ab2 = new bg.ArrayBuffer()
> ab2 instanceof ArrayBuffer
false
> ab2 instanceof bg.ArrayBuffer
true

So that's annoying.  Invert the check in writeUTF8 so it tests the
argument as a string first, otherwise assume it's an ArrayBuffer or
a typed array or compatible.

Bug: chromium:930171
Change-Id: Ib6dacaa2b6ecf8bd7535d5d938863d4e78c1d1bd
Reviewed-on: https://chromium-review.googlesource.com/c/1466381
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>