nassh: SSHAgentRelay: pull out message size parsing
We continue to queue data until we have an entire message before
parsing it entirely.  The code currently will keep rescanning the
first 4 bytes (message length) until the entire message is queued.
Lets change the code a bit to extract & cache the length once it's
available.  This adds a few more lines of code, but it will help as
we migrate to ArrayBuffers.

Change-Id: Ied6fbcc2be680d6bc9769b1cef238ebac9f7571d
Reviewed-on: https://chromium-review.googlesource.com/c/1470528
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>