nassh: SSHAgent: pull out message size parsing
We continue to queue data until we have an entire message before
parsing it entirely.  The code currently will keep rescanning the
first 4 bytes (message length) until the entire message is queued.
Lets change the code a bit to cache the length once it's available.
This adds a few more lines of code, but it will help as we migrate
to ArrayBuffers.

Change-Id: I3bf499ca6e697c14b23f8727c5b2306a94d5ead0
Reviewed-on: https://chromium-review.googlesource.com/c/1472391
Reviewed-by: Fabian Henneke <fabian.henneke@gmail.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>