nassh: Add support for Ed25519 via OpenPGP applet
Adds support for SSH authentication using the Ed25519 curve on an
OpenPGP-enabled smart card to the 'gsc' agent backend.

The commit is structured in a way that allows support for other EdDSA
curves, such as Ed448, to be added simply by appending their OIDs and
public key prefixes to the object nassh.agent.messages.OidToCurveInfo.

Change-Id: I53ea12072fdee75f16d2dbc7ef0b3c3d472d35ac
Reviewed-on: https://chromium-review.googlesource.com/c/1474932
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>