nassh: Only offer PIN caching if using agent forwarding
The cache for smart card PINs currently only stores the PIN for the
duration of an SSH session. If agent forwarding is not used, the PIN is
entered only when the session is established, so it does not make sense
to offer caching in that situation. With this commit, PIN caching is
only offered when agent forwarding is being used.

Change-Id: I1a58c9fec190a5405d9daffaedcaa1fce1ee2892
Reviewed-on: https://chromium-review.googlesource.com/c/1474937
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>