nassh: crosh: don't assume ids are pids (numbers)
We've treated the id from the chrome.terminalPrivate APIs as an opaque
value in all places except one: we use -1 for errors and uninitialized
sessions.  In practice this doesn't matter as the only time we check -1
from the Chrome side is when we open the initial session which seems to
be unnecessary (as the value is actually undefined).  Otherwise we are
checking the -1 value we set ourselves and then passing it down to the
Chrome layers directly.

We want to change the ids used by chrome.terminalPrivate to be a random
string, so rework the crosh code to use null as a sentinel value so we
avoid any possible confusion.  We retain the initial -1 error checking
for backwards compatibility.

Bug: chromium:352465
Test: loading crosh on existing system works, and on new id/string system
Change-Id: Iffc56e3a30c1a57c2d8ef2d115a12fa9dae675f9
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1478417
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>