nassh: sftp: change getData helper to return typed arrays
In preparation for converting to ArrayBuffers, change the getData
method to return a typed array.  This allows us to make a lot of
mechanical changes (like tests) before making functional changes
to convert the underlying code to use ArrayBuffers everywhere.

This makes the getString method a little more expensive by running
the string through stringToCodeUnitArray and then back through the
codeUnitArrayToString helper.  We'll clean this up in the future,
but for now, it's not a big deal as getString is only used for a
few short strings (on the order of <10 bytes).

Change-Id: I7e801dba9240dcd8d6c5149801ce08b6ce7b64c4
Reviewed-on: https://chromium-review.googlesource.com/c/1482312
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>