nassh: sftp: pull out message size parsing
We continue to queue data until we have an entire message before
parsing it entirely.  The code currently will keep rescanning the
first 4 bytes (message length) until the entire message is queued.
Lets change the code a bit to cache the length once it's available.
This adds a few more lines of code, but it will help as we migrate
to ArrayBuffers.

Change-Id: I69d723da355dc6f28e76ce439c5e5d34790d130e
Reviewed-on: https://chromium-review.googlesource.com/c/1482314
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>