nassh: sftp: simplify the buffer parsing logic
The loop that that repeatedly calls parseBuffer does more checking
than it needs to.  parseBuffer already checks the buffer length and
returns false if there's not enough data available.  So we can call
it as the main loop.

Lets internalize the parseBuffer function because nothing else needs
access to it.

We drop the catch/logging logic as we already check the lengths in
the lower layers.  If there's logic errors lurking about, we should
be making them a bit more visible too.  At this point, we don't have
any known/documented failures.

Change-Id: I3f7c55f057c5a64c98cd9335105f752fa515e8ad
Reviewed-on: https://chromium-review.googlesource.com/c/1482315
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>