nassh: crosh: fix opening in window from a tab
The lib.f.openWindow API works if the current session is already in a
window.  If it's in a tab, then Chrome (for some reason) blocks using
chrome=no in the window.open call.  For crosh, we can assume we have
access to Chrome APIs as this only runs on CrOS, so use chrome.windows
to get guarantee opening as a new window.

Change-Id: I644eb12dcd5fdf2c0aa3e8a3ced6bbad4406f8f1
Reviewed-on: https://chromium-review.googlesource.com/c/1495414
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>