nassh: crosh: make it easier to open as a window
Support the openas parameter in the URI like we do with nassh already.
We switch to the standard URLSearchParams instead of lib.f.parseQuery
as we don't need to support arrays here.

Further, if we aren't open as a window already, then issue the same
pro tip message as nassh, but instead of linking to the FAQ, create a
link that automatically opens a new window users can click on.

Change-Id: Ie61482f1ef6f8aaef298a3edacbd278a709cc8d7
Reviewed-on: https://chromium-review.googlesource.com/c/1495419
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>