hterm: tests: move state cleanup to postambles
Rather than relying on test's preamble to possibly clear any state in
the document by nuking everything, use postamble to undo any setup we
know was run.  This makes the tests generally more reliable, and it
makes it easier to switch to mocha's test runner.

The fullscreen test needs a little more rework as it creates a terminal
independent of the terminal already created in setup.  It doesn't need
to reuse the existing state, so stop clobbering the attached fields.

Change-Id: Ifa0ac96cc4f7c8d247ac969de6c3d224062c2b8b
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1628511
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>