nassh: rework how we tear down the plugin
Our current (main) exit dance looks like:
- NaCl sends exit(code) message to JS
- JS sends onExitAcknowledge message to NaCl
- JS shows exit message (in parallel)
- NaCl exits (in parallel)
- The plugin stays resident until we reconnect

This often runs fine.  However, we seem to frequently run into
misbehavior with how we exit the NaCl runtime.  This manifests as
the NaCl runtime crashing which in turn can trigger messages from
the host OS that a program crashed.  We've debugged this a bit
(see the referenced bug), but at this point, the NaCl runtime is
EOL and never going to improve.

Rather than telling the plugin to exit itself, we can hard delete
it from the DOM from the JS side.  This seems to trigger the NaCl
runtime to hard exit without triggering the crash flow.  By the
time we want to exit, the plugin has finished flushing and closing
all of its own state, so we shouldn't be losing any data.

We also change the flow to remove the plugin whenever the JS code
calls the instance's exit method.  In the past, we wouldn't signal
the plugin or otherwise try to close it, so in theory it might be
running still, even if we've half-closed the overall JS state.

Bug: chromium:710252
Change-Id: I2677531ff9a1130aad55dbb637ec631fc5b6fbd2
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1514031
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>