nassh: allow agent message to accept Arrays
The sshfe code passes Arrays to the agent layers as it's reading data
via Chrome's messaging APIs, and those only support Array types.  Fix
the agent layers to accept this so it just works -- the code is free
for us as the Uint8Array constructor accepts a range of inputs.

Change-Id: I65b0cd60d5a2ea9b589bdfc1403d460252a0677d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1526923
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>