nassh: fsp: handle symlink deletion
This isn't useful just yet as we filter out all symlinks from the FSP
layer, but it'll matter after we start exposing symlinks to it.  Since
we'll bubble up symlinks to directories as directories, we need to make
sure we delete symlinks as symlinks even if the FSP layer thinks they
are directories.

This might be confusing to users if they think they're deleting the
underlying directory, but since we haven't exposed symlinks previously,
this isn't a regression.  Let's see how users handle it in practice.

Bug: chromium:714212
Change-Id: I8d8be02965d38bad1e07a7a27034f8ba99bead19
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1528002
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>