nassh: identity: list all id_* files in dropdown
We don't actually need the .pub half of the key, so requiring it
inorder to display the private half is unnecessary.  Tweak the
search algorithm to list all id_* files for now as that should
cover a significant number of users.

We switch to a Set to keep the code simple.  If we find both the
id_xxx and id_xxx.pub, we'll add each in the diff code paths.

This also has the nice side effect of keeping the identity list
sorted vs today's "whatever order the filesystem returns".

Bug: chromium:668754
Change-Id: Ie9ac7c139485bab6626d559fd2fb19579fc7db38
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1528008
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>