nassh: identity: allow resetting back to the default
The current identity logic will ignore the default value when saving
which means you can select a non-default identity, but you can never
go back to the default value.  Attempts to change it are ignored.  To
pull this off, we need to leverage the .value field as an independent
setting from the user-visible .textContent field.

While in here, delete an identity event listener for an event that
is never fired -- the select event is for arbitrary text inputs.

Bug: chromium:668754
Change-Id: I3dcaeedb155c6a43f4322337767170fa7c3bce96
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1528017
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>