nassh: identity: improve the import process
The current importer will load multiple files and resync the UI after
every single one.  This isn't normally a problem as people tend to not
import more than one file at a time, but it's still wrong.

Inline the nassh.importFiles logic here as it's the only user (and has
always been the only user).  Then turn each import into a Promise so we
can resolve all of them and then refresh the UI after they're all done.

This also allows us to fix a long standing bug: while we'd update the
UI to make it look like we selected the new imported identity, we never
save that setting, so we never actually select it for real.  With this
code restructured, we can add a single save call to the right place.

Bug: chromium:668754
Change-Id: I91a90b4b8bc6516c85d415f25e1eea0723ff4569
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1528018
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>