nassh: identity: move key files to /.ssh/identity/
Dumping imported keys into /.ssh/, while matching normal the Unix flow,
is confusing to users, and makes it hard to reliably update the UI, and
to avoid file collisions against normal internal ssh files.

Historically we've forced people to import both the private & public key
halves so we'd have a way of finding keys that end with ".pub" to display
in the UI.  We did this to avoid listing files like "known_hosts".  But
in reality, the ".pub" half isn't necessary, and users find it confusing.

It also turns out that users will use whatever filenames they feel like.
So we can't rely on any standard like "id_*" as ssh itself uses.

By moving everything under /.ssh/identity/, we create a reserved space
that we know ssh doesn't use and thus everything in there is safe.  We
still ignore *.pub for historical reasons, and in case people import it
still and get confused when it doesn't work.

This should be backwards compatible as we don't move any existing key
imports, nor do we update any preferences.  Since we can set the user
visible string in the dropdown different from the actual value, we can
keep the basename & full path separate to hide the identity/ subdir.

This also helps out in another way: ssh will search certain id_* files
implicitly even when the Identity box is not selected.  So if you had a
"id_rsa" key but didn't explicitly pick, ssh will try to use it.  We
aren't breaking backwards compat (existing files remain in place), but
when people import the keys in new systems, they will need to explicitly
choose the identity in the drop down.  Doesn't seem like a big deal, but
we can wait for user feedback.

Bug: chromium:668754
Change-Id: I7ebb529cd26fce46d6369a773c7ef2369b89981d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1536726
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>