libdot: test: support async preambles
A bunch of hterm's tests create a new terminal instance which requires
async initialization.  Currently all our tests have been racing and we
happen to work, but we should really be testing a fully setup instance
to properly reflect the real state.  This is also required as we switch
over to mocha.

Change-Id: Ia03e1597fefc943d4a928f6ea7983594026efce5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1628513
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>