ssh_client: overhaul toolchain management
We want to start adding wasm support to the builds, but the current
logic assumes the target is always pnacl.  Lets rework how toolchains
are selected so we can make it into a command line option.

We formalize the interface a bit by creating an object with standard
fields that packages actually care about.  We also push the settings
into the environment for all source functions so builds can find it
via standard methods like $CC.

We can collapse the /usr subdir into just the sysroot.  None of the
compilers care about the split-/usr behavior, and wasm fails even
more if you try to do it that way.

Finally, we move builds under their respective toolchains.  So instead
of output/build/$P/work/, we have output/build/$TOOLCHAIN/$P/work/.
This will let us build pnacl & wasm side-by-side which we'll want in
order to ship them in parallel for early testing.

Change-Id: I22ce6ffea63a97365ba318e0e624cb865dcbca3a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1805544
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>