ssh_client: ldns: roll in some patches
Upstream hasn't made a new release in a while, so we have to pull in
some fixes ourselves.  The first two are for some security issues,
while the latter two are for build related failures depending on the
OpenSSL configuration.  We want to remove unused features from our
build of OpenSSL which trips these failures.

Change-Id: I186631360963c3fe29fa8cdaf4cb145d0ba1416f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1647274
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>