hterm: tests: create fresh scrollport for each test
The current scrollport tests create a single scrollport for all of
them, then try to set it back to a default state before running the
next test.  This seems to work in the current framework, but when
moving to mocha, some assumptions we have about overall state of
the DOM causes them to fail.  It's probably bad for us to try and
run tests on dirty scrollport objects anyways when we don't offer
any APIs to reset it back to initial state (we just scroll up).
Change the code to create a new scrollport for every test instead.

Change-Id: Iea16c4876e0b56a6b6643b5d934e14ccba40c43a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1649697
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>