libdot: concat: explicitly set file encoding to UTF-8
Make sure we always read/write files as UTF-8.  We've been relying on
the default environment encoding which is normally UTF-8, but if it's
not, then we can crash when implicitly converting.  Most of these are
not an issue because the content is restricted to ASCII, but our output
always contains UTF-8.

Change-Id: I70f6fed6ffa2e3075c28d6a25dbbad875b825c43
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1652242
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>