libdot/hterm/nassh: convert load_tests.sh to python
This allows us to standardize the interface across all of them and
kill off one more source of shell scripts in the tree.

For hterm/nassh, we have to create little python modules to help
bootstrap their logic (since this is the first python either has
seen).  The style is cribbed from the existing ssh_client.py.

Change-Id: I2478f598c574cbff8ca1168e3a366b7a5031a964
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1652244
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>