nassh: workaround fd reopen ack race
Newer OpenSSH has uncovered a race with our fds:
- Plugin->JS: openFile 101 /dev/tty
- JS->Plugin: onOpenFile 101
- Plugin->JS: write 101 1 (sent count increased by 1)
- Plugin->JS: close 101
- JS->Plugin: onClose 101
- Plugin re-opens file which resets ack/sent count to 0
- Plugin->JS: openFile 101 /dev/tty
- JS->Plugin: onWriteAcknowledge 101 1
  - This is bad -- we're acking a write on a fd that was closed.
- Plugin now has sent=0 ack=1
- JS->Plugin: onOpenFile 101

At this point, the counters are out of sync which makes math on them
underflow which makes it look like we've written more data than the
window allows, so the plugin refuses to send out anymore data until
the other side acks enough pending data.

Normally this isn't an issue because OpenSSH doesn't close & re-open
the same fd quickly.  It opens a fd and keeps it until it exits.  We
could fix this a few different ways that would require overhauling,
so for now lets just avoid sending acks on streams that are closed.
This seems good enough for this edge case with /dev/tty.  I expect
for our migration to WASM that the plugin & communication is going
to be completely rewritten anyways, so I don't want to spend too
much time chasing this down.

Change-Id: If2da042f694c01355c0a2577f9a976f99d97ab89
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1652577
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>