nassh: run tests against local node server
Using ES6 modules requires using <script> tags with type=module, and
that in turn requires loading files not from file:// URIs (since file
isn't in the browser's list of supported URIs).  So our current method
of pointing a browser to the test page via file:// no longer works.
There doesn't seem to be anyway to bypass this restriction, so we use
the http-server module to spin up a local static webserver to run the
tests against instead.

We're only using ES6 modules in nassh currently, so we only turn this
on for it.  We can still run libdot/hterm in the file:// method.

Change-Id: I7efcf287d5a9c0e1488d5bf50827c60d8a0b5e04
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1687490
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>