nassh: write our own background page
We've been relying on Chrome to generate a background page for us on
the fly using the scripts array, but that doesn't work if you want to
use ES6 modules.  In order to add the required type=module attribute,
we have to write our own background page and list out the scripts.

In practice, this shouldn't change anything.  For the extension, we
include some extra scripts we weren't before to avoid maintaining a
second background page, but this doesn't seem like a big deal when it
is usually idle & non-existent.

Change-Id: I0fe3fe89f90bf5d57b03698c2c6e0933bd0aee98
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1690393
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>