nassh: implement new corp-ssh-v4@google.com protocol
This implements the new protocol Google is migrating to (GRUE).  The
old relay protocol code is kept separate to avoid breaking any users
outside of Google, and to basically keep it frozen indefinitely.

There was a v3 internally during development & deployment, but all
clients & servers are at v4, so that's what we're releasing with.

Change-Id: I2d668642ba093628bb8464563d56c80d7167baff
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1870095
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>