nassh: add ssh_client license details too [reland]
We roll a lot of software from OpenSSH, so make sure we include it.
We have to fix some non-UTF8 encoding issues in their files.

While we're here, polish the UI a little to match chrome://credits
by hiding all the data by default.

The crosh ebuild has been fixed, so we can reland this.

Change-Id: Ia2fb3c2009eb78423f116f3216773dd7f200674a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1734045
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>