kokoro: run under /libapps dir
When we run in docker, we build inside of a full distro.  Instead of
mounting the various libapps dirs directly into the filesystem root,
use a /libapps/ subdir.  This simplifies the run command as we bind
in all of the git checkout, and it avoids issues with creating new
dirs on the fly in the / directory.

Change-Id: I8400a7767937f2c9408bed72d9380f05b5398f84
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1727431
Reviewed-by: Max Lin <linmax@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>