nassh: use absolute html paths everywhere
We've been using absolute /html/ paths in most places already, but it
looks like Chrome changed behavior on us and our use of relative html/
paths is breaking in some situations.  Just change the few places over
to absolute paths to fix.

Specifically, if you try to do File Manager -> Add new service -> SSH,
it opens /html/html/nassh.html.  My guess is because it's using the bg
page context which is already /html/.

This might be a Chrome bug, but not a big deal to us.

Change-Id: Iff07005fc27f57bf7e9a6ed63a1ca54c4c99b1e5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1764958
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>