nassh: Allow non-numeric PIV PINs to be sent to the card
While the PIV spec requires PINs to be numeric, cards and middlewares
are more permissive and many users rely on this larger PIN space.

To accomodate them, we drop the requirement that all characters of the
PIN be numeric while keeping the length requirements.

The spec explicitly demands that PIV card applications require the PIN
to be numeric before presenting it to the card, but it is extremely
unlikely that cards would not do their own validation.

Note: Reconsider this commit upon user reports of breakages related to
PIN entry.

Change-Id: Ia44ffb086d4d8f8fb4fb98cceaa6edd4a0339b0a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1763541
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>