hterm: pass commandName in Command constructor
hterm.Terminal.prototype.runCommandClass is modified to
take commandName as a parameter which it inserts into
argv.commandName when creating the command class.
It also changes string field argString to be args
which is an array of strings.

This improves the crosh code to make commandName a normal
field of the class.  Previously the code was modifying
Crosh.prototype.commandName to override commandName
from url params.

The pattern for the CommandClass to take a commandName
will also be used in terminal code which will use a
default commandName of vmshell.

Nassh code is modified to pass args as an array and
expect to have argv_.args rather than argv_.argString.

Change-Id: Ib9d0a5689a415de328f1718d57361ea258a58afe
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1771679
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Joel Hockey <joelhockey@chromium.org>