ssh_client: improve --host/--build handling
Pull the hardcoded --host=nacl settings out into the toolchain so we
can change it based on the target.  We also need to pass down --build
as configure scripts will default --build to --host which is not what
we want.  This happened to work with nacl, but doesn't with wasm.

Change-Id: I6cea7fba16dcccb47592281bc12385d8c505c5dc
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1831004
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>