libdot: Added lib.assert and lib.notNull to help JS compilation and tests
The JS closure compiler valiates that parameters passed
to each function match the required types.  It is a common
pattern to have a member such as a string which is null
at object creation time, but becomes populated before
it is subsequently used as a parameter to a function
which is expects a non-null string.

The closure compiler will fail unless it can deduce
that the nullable value is guaranteed to be non-null
when it is used.  This requires us to guard all uses of
a field behind a null check.

lib.notNull(foo) is a shorthand way to express
to the compiler that we are confident that the
field is not null when it is used, and is a useful
guide to developers.  It has annotations to assert
that the condition is truthy.

lib.notNull uses lib.assert which is a nop in
production, but uses chai.assert when it detects
the presence of chai in tests.

Bug: 998928
Change-Id: Iad4807339f4cc922a515870bc38efdedd11d4e1a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1816024
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Joel Hockey <joelhockey@chromium.org>