kokoro: fix ownership of artifacts output
Since the container runs as root, the artifacts output dir is owned by
root which means code (like the CI) outside of the container is unable
to modify files in there.  We need to generate a file or two, so make
sure we reset the uid/gid to the accounts outside of the container.

Change-Id: Ifa2e16abbd32011915fd3b98eca707e78742f19b
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1827684
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>