kokoro: slim down container
Since Debian buster doesn't have python3.6, the request to install it
ends up acting as a regex and pulling in a ton of packages we don't
want.  Switching that to python3.7 avoids the problem even if we are
now testing against a version newer than CrOS.

This also means the container size is cut in ~half which should help
speed up the CI in general.

Change-Id: I68ba839b4c34370e8ba9f90ce00eb539c60dec2a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1829808
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>