libdot: fix portability issues with unpack
When unpacking paths, we've been passing the absolute path to tar.
On Linux systems, this largely works fine.  But if the path contains
a colon (which comes up frequently under Windows), tar treats it as
a remote specification and fails trying to connect.  We can mitigate
this by passing tar a path relative to its cwd which we control.

Change-Id: I7b8a0caef7115cc18275fa711eca64b178d773a6
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1831001
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>