terminal: Set startup_id and send to chrome
Terminal relies on having a unique ID for each tab so that
text sent from a shell process is only processed by the
intended terminal.  This ID is generated when the vsh process
is created and sent in response to openTerminalProcess.

We have made a recent change to send startup text to
the terminal while crostini is starting and before
the ID has been set.

Terminal will now generate a random number to use as the
ID and send it in openTerminalProcess args.  This ID is
used for startup text until openTerminalProcess callback
is invoked and the process ID is then used.

Change-Id: I0f5b506d4e70bb210e4080bf04c6d11df894f71e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1875864
Reviewed-by: Julian Watson <juwa@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>