nassh: rewrite release generation
Rewrite the existing mkzip shell script logic into a new mkdist
Python script.  This isn't a straight port where we retain all
the functionality of the old code as that was a lot more general
than we needed (only nassh does this), and it wasn't worth the
overall effort.

The old method had us running 3 scripts to produce the release
artifacts in here.  The new method just generates them all in
one step as there's little overhead here (O(seconds)).

We leave the shell scripts in place for now so we can update the
CrOS ebuilds to the new code first.

Change-Id: I753f252d7f10cf28fb74dc34f3839f8d25fb2574
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1882412
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>