wassh: implement blocking syscalls with memory writeback.
Syscall handling takes place on the main therad. The worker thread,
which holds the WASM runtime and its associated memory, can now
postMessage to the main thread's syscall handler and block its own
execution until the syscall handler is complete.

Memory updates are still a problem. This may be fixed when the WASM
runtime supports SharedArrayBuffer instances as its backing store for
Webassembly.Memory types, but until then we're stuck with having to
perform manual memory writebacks. In order to try and eliminate
copying overhead SharedArrayBuffers are used for each memory region
before passing to the syscall handler.

Not all syscall handlers will be able to write directly to shared
memory, notably the Crypto.* family of APIs will not work directly on
shared memory, mandating a further copy operation.

Change-Id: Ief22f80da635444dcd5e94ab5a6ddf58c16526db
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1975311
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>