nassh: Add loadMessages, reuse hterm messageManager
Added nassh.loadMessages for nassh/crosh to be able to load
messages in an environment which does not support chrome.i18n.
This will be used when crosh runs in chrome://terminal in
order to gzip resources.

Loading messages like this requires the use of a
MessageManager class, so we can reuse hterm.messageManager
once it is reconfigured with useCrlf on.

Bug: 1019021
Bug: 1022250
Change-Id: I96ffc54ef9052337ad342a3b08c43a120ea1a77d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1918864
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>