nassh: fix up new --welcome setting
Add missing documentation and checking in the connection dialog.

Fix option checking in startup to avoid breaking SFTP mounts (where
the args member is not set) and to properly handle the different
command line parsing variants (e.g. coalescing --welcome --no-welcome
and such).

Move the clearing to before we start relay parsing so we don't throw
that all away (and because it's not part of the "welcome screen").

Switch the clearing logic to use the clearHome API rather than ad-hoc
terminal escapes as it's much easier to read & understand.

Change-Id: I363ca03d2a3e7374eeb5fe5fe7ea32d375062430
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1946388
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>