nassh: drop connect dialog from web_accessible_resources
There wasn't any documentation when this was added to the manifest.
It doesn't seem to be needed as the dialog is only opened by the
extension itself, and we don't need it for relay flows.  Drop it as
we don't want to list any paths in web_accessible_resources unless
needed as that allows people to link/embed the page.

Change-Id: Ia8c5c42095d6442f5b24357ae643eca22ff1dd71
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1965962
Reviewed-by: Andrew Stone <ajws@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>