nassh: add crosh & socket APIs
We add crosh access so we can replace the app variant entirely.
We've been whitelisted in M81, but listing the API shouldn't
block loading in older versions.

We add the socket APIs for the upcoming WASM work.  We only add
this to the extension as we'll keep the app on NaCl until it dies.

We need UDP to query DNS directly ourselves so we can process the
extra records (like SSHFP).
We need TCP (client) in order to connect to servers of course.
We need TCP (server) in order to support local port forwarding.
It's not a super common use case, but people def use it.

Bug: 1034178
Change-Id: I4927eea408d493c2da230ff65ea00f23da2a1c7c
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1967223
Reviewed-by: Andrew Stone <ajws@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>