nassh: start encouraging people to move to the extension
The Chrome App (and its tech stack) are deprecated.  The extension
is on a path to migrate to supported technology for all platforms,
so lets start driving people to migrate off the app.

We'll display a notification to users on R83+.  All of the things
we need should be in Chrome R81+, but R83 will make sure that R81
is in stable when we start getting people to migrate.

Change-Id: Ic45849947695ca71a9e19b3fc441ddcd477c789e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1968222
Reviewed-by: Andrew Stone <ajws@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>