wassh: refactor syscall handlers
This starts the generic WASI JavaScript Bindings project using the
existing wassh code as a basis & inspiration.  All the code in the
new project is meant to be program agnostic (i.e. have nothing in
it specific to wassh) while everything specific to us still lives
under wassh/.

More work is needed to flesh things out further (like implementing
more syscall entry points), and as the wassh logic matures, we can
consider what could be moved out as a general framework for others
to re-use, but this is a good start to build on without having to
implement the entire world first.

Change-Id: Iaa3fc41be747b83e02213fff30845350c200d333
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2012260
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Andrew Stone <ajws@google.com>