wasi-js-bindings: clean up jsdoc/closure logic
Fill out the jsdoc comments and add closure externs.  This enables us
to start running our linting tools to check code quality/behavior.  We
delayed this at first to get a working poc, but now the code is getting
complicated enough that the linters are catching errors.  There should
be no functional changes in here.

Change-Id: Ie4be74d772e8956414e342027833d077b624a873
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2484162
Reviewed-by: Collin Moore <moorect@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>