nassh: sshfe: make parsing a little more robust
We had a generic catch() in place to handle any errors, but this
only worked for code inside of a promise chain.  The initial get
challenge logic ran outside of any promises which meant errors in
it (e.g. user typoing the proxy host) would throw exceptions that
no code caught.  Move it all into a promise so the existing catch
logic also handles it.

Change-Id: I85b7f70e600ff1ac58661b8fa66352237ca980ba
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2019773
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>