nassh: have Ctrl-Shift-N open connect dialog for extension sessions
When connecting to URIs, we want Ctrl-Shift-N to open the same
connection rather than the new connection dialog.  When connecting
to saved sites via the UI though, we want Ctrl-Shift-N to open the
connection dialog.  The app does this, and the extension does this
if you open the connect dialog initially, but the extension doesn't
do this if you use the popup connection dialog.

Have the popup UI signal to the new window that it should go back
to the connection dialog when people use Ctrl-Shift-N.

We might make this into a pref in the future, but for now, let's
keep the UX the same between the app & ext variants.

Bug: chromium:1045581
Change-Id: I7a04dbd225f9c9cc964caf7ef7fa0582717bc99c
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2023275
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>