hterm: use \u escapes to workaround builder troubles
The builders are producing a corrupted concat hterm file with the
UTF-8 content.  While we debug that, switch the file to \u escapes
so we can get a release out.  I'll leave a TODO in here with the
UTF-8 content so we can tell if the builders are still corrupting.

Change-Id: I80ce96f640c8224e11229332b16d497879eba7b5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2030064
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>