terminal: Set $i18n{themeColor} in terminal.html
Currently we set page theme after a page loads
after we fetch prefs.  This causes flickering of the
frame and page bg since the page is first set to
the default theme from app manifest.

In order to render correctly from the start,
we will fetch the theme from prefs and substitute
it into terminal.html for meta theme and
css background.

Bug: 1033339
Change-Id: Ic85f9c7270bb4587db6fefdd0244067b4f3f3302
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2056189
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jason Lin <lxj@google.com>