terminal: store preference colors in HSL
The preference color picker operates on the HSL color space. Before this
CL, we stored the color value as RGB values, and this caused a bug that
the hue slider jumped around when you adjusted the saturation/lightness.
The reason was that the color picker would need to convert the HSL color
to RGB and then convert back again, and the conversion is lossy.

This CL solves the problem by always storing color in HSL.

Bug: 1050921
Change-Id: I3916afae8c0bd3a10782fbf0047b4881c4cd9d6e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2071677
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Julian Watson <juwa@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>