terminal: add ansi colors to settings
The <terminal-settings-ansi-colors> element
binds to the 'color-palette-overrides' preference
and groups 16 <terminal-colorpicker> elements
in 2 rows of 8.

The colorpicker has been modified to split out
<terminal-colorpicker> which does not bind to
a preference, from <terminal-settings-colorpicker>
which does.

A new inputInDialog property is added to
indicate that the hex input field should be
shown in the dialog rather than next to the
swatch.

Bug: 1057507
Change-Id: I22923e7d6daf3303ff2f339fd9aedb1a1c7c8b72
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2084319
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: Joel Hockey <joelhockey@chromium.org>