nassh: mkzip: start building versions for stable rollback
If a stable upgrade needs to be rolledback, the CWS does not provide
a great way to handle that: it requires the version field always be
newer than the latest uploaded version.  To workaround that, we can
create parallel versions with a slightly higher version number.

So if we release 0.23 today, we'll build an identical extension but
with the version set to 0.24.0.1.  If we release 0.24 tomorrow and
we want to revert that back to 0.23, we have the 0.24.0.1 version
already built & ready to be deployed.  It will be exactly the same
as 0.23, but the version number is larger than 0.24 so CWS will
allow it.

The UX will be a bit confusing as 0.24 will be nothing like the
"newer" 0.24.0.1.  We can add some messaging to explain things, or
maybe add a FAQ, or something.  But we'll do that in a follow up.

Change-Id: I30a5ade59b4cbdd9abf20e7d282c189551daca4a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2086896
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>