nassh: relay corpv4/sshfe: fix write ack confusion
The plugin API expects the write count acks to be the total byte
count for the current session, not the size of the most recent
write request that it made.  This can cause the writes to queue
up if a lot of small requests are made in succession as the ack
counters get out of sync, but the issue was hidden when using a
lot of large writes as those would also happen to set the ack
state correctly to allow data to keep flowing.

Bug: chromium:1046834
Change-Id: Ie7eb046331a0285eea9fc64a822e5098ba88da5f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2090195
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>