nassh: retain the original destination in the location
We use the location.hash to store the encoded connection info for
where we're connecting to.  This is used partly for users to see
what site a page is connected to, and for reconnecting.

When parsing a ssh:// URI, we would partially mangle it by
stripping off the schema prefixes before decoding, but then we
would store the partially stripped (but not decoded) result in
the hash.  If we then try to reconnect, we wouldn't decode the
settings since the schema prefixes were removed leading to an
error when using the encoded form directly.

Lets have the client store the original destination form in
the hash so that we can always reconnect with the same value.

Bug: b/151159072
Change-Id: I56f4414830fefec0852d5c605c77b1f4a223dbc5
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2101499
Reviewed-by: Collin Moore <moorect@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>