nassh: nasftp: make command loop more robust with rejected promises
If the subcommand rejects promises directly, or throws exceptions, we
have handlers in place to catch & process them correctly.  But if the
subcommand delays execution to event handlers or async callbacks, and
those reject promises or trigger exceptions, Chrome cannot chain that
to the right handling context, and the CLI hangs waiting for the call
to finish processing.

Register with the unhandledrejection event so we can catch those async
failures too, and then diagnose/recover correctly.

Change-Id: I4cf499e374fb537d73ac309f1a85ca6175a180d4
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2112942
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>