nassh: nasftp: ls: fix fake file listings
When the cwd is the default (i.e. "."), then the constructed "full
path" will start with "./" all the time.

When doing `ls` on a non-directory, we fake our own results using a
stat call.  But we use the fully constructed path as the filename.

When we go to display the gathered file entries, we skip all the
ones that start with "." unless the -a flag is specified.

Taken altogether, our attempts to make `ls` work on files doesn't
actually work by default.  Fix it by using basenames in the faked
results.  This matches the default listing behavior better anyways.

Change-Id: I2da58fdf3b4889a893e9e00d275118aeae767572
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2112941
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>